Change password
Users and admins can now change passwords. An email is sent to the user
whose password was changed. Admins will see the "old password" field
when changing the password for a user but can enter anything in this
field as long as it passes basic password validation. The old password
is not specifically validated against the user for admins.