Grant Award
Awards can be granted from the awards page.  Form needs to be protected for adminlevel.  Form needs styling.  Would be nice if was in a modal screen and out of the way.
Ability of delete awards needs to be added to the tables.
Repaired Profile controller.  Parts missing during last merge