Article editing complete
New way of handling article editing. This now works much better than the
old way of having the article editing setup in the individual
controllers. See the edit_hub function in the Articles controller for an
example of what needs to be done to edit articles of other types.

Fixes #121