Datatables for everything.
See #235 for discussion.

This only implements the new setup with the admin/airports. It should
work out of the box to return all public properties for other models,
but if you want to limit what is shown in the table you will have to set
the columns in the appropriate model to match the columns displayed in
the table. This is how it is done in the Airports model example.